[RBTree] Decl-Impl Separation
